requirement.txt
Django==5.1.7
pytz==2021.1
sqlparse==0.4.1
starlette == 0.26.1
psycopg2-binary
gunicorn == 20.1.0
asgiref
djangorestframework
django-crispy-forms
